requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pandas
lightgbm
scikit-learn
joblib
numpy
